+++ include/ftk/mesh/simplicial_mpas_2d_mesh.hh
@@ -6,7 +6,7 @@
 
 namespace ftk {
 
-template <typename I=int, typename F=double>
+template <typename I=int128_t, typename F=double>
 struct simplicial_mpas_2d_mesh : public simplicial_unstructured_2d_mesh<I, F> {
   simplicial_mpas_2d_mesh(
       const ndarray<F>& coords,
